refactor: state 객체 별도 파일로 분리
